+++ social-seo-metatags.php
@@ -105,6 +105,53 @@ class SocialSEOMetaTagsPlugin extends Plugin
 
     //Set new meta to the page
     $page->metadata($meta);
+
+    // Build breadcrumb
+    if ($this->config->get('plugins.social-seo-metatags.seo.breadcrumb')) {
+      $this->buildBreadcrumb();
+    }
+  }
+
+  /**
+   * Build JSON-LD breadcrumb
+   *
+   * @doc https://developers.google.com/search/docs/advanced/structured-data/breadcrumb
+   */
+  private function buildBreadcrumb() {
+    /** @var $page Page */
+    $page = $this->grav["page"];
+    /** @var @var $assets \Grav\Common\Assets */
+    $assets = $this->grav['assets'];
+
+    $inline = '{
+  "@context": "https://schema.org",
+  "@type": "BreadcrumbList",
+  "itemListElement": [';
+
+    // Current page
+    $inline .= '{
+    "@type": "ListItem",
+    "position": 1,
+    "name": "' . $page->title() . '",
+    "item": "'. Utils::url($page->url(), true) . '"
+  }';
+
+    // Children
+    $index = 2;
+    foreach ($page->children() as $child) {
+      $inline .= '{
+    "@type": "ListItem",
+    "position": "'. strval($index) .'",
+    "name": "' . $child->title() . '",
+    "item": "'. Utils::url($child->url(), true) . '"
+  }';
+      $index++;
+    }
+    $inline .= ']
+}';
+
+    $assets->addInlineJs($inline, null, 'bottom', 'application/ld+json');
+
   }
 
   private function getSEOMetatags($meta)